source: address review findings (redaction, nil guards, test quality)
Post-review hardening of the source coverage work:

- location: mask password-only userinfo ("scheme://:pw@host") in
  redactBestEffort. The username regex group is now optional, so an
  unparseable or unknown-scheme location no longer leaks the password
  via Short/Redact or error messages.
- metadata: skip nil column entries in Table.Column and Table.PKCols,
  matching the nil-element discipline already applied to Source.Table.
- tests: assert LogValue's emitted attrs (catalog/schema) and location
  redaction instead of a no-op smoke log; add nil-column and
  password-only-userinfo cases; fold coverage_test.go into the
  conventional source_test.go / handle_test.go files and merge the
  duplicate Contains coverage into TestContains.
